refactor(managesieve): rebase on new oidc token validation
