+++ protocols/managesieve/src/main/java/org/apache/james/managesieve/api/Session.java
@@ -23,8 +23,8 @@ package org.apache.james.managesieve.api;
 import java.util.Optional;
 
 import org.apache.james.core.Username;
+import org.apache.james.jwt.OidcSASLConfiguration;
 import org.apache.james.managesieve.api.commands.Authenticate;
-import org.apache.james.protocols.api.OidcSASLConfiguration;
 
 public interface Session {
