+++ server/protocols/protocols-managesieve/src/test/java/org/apache/james/managesieveserver/ManageSieveClient.java
@@ -92,8 +92,8 @@ public class ManageSieveClient extends SocketClient {
 
                 response = new ServerResponse(responseType, responseCode, explanation, lines);
             } else if (tokens[0].equals("+")) {
-                Optional explanation = Optional.of(tokens[1].substring(1, tokens[1].length() - 1));
-                response = new ServerResponse(ResponseType.CONTINUATION, Optional.empty(), explanation, new ArrayList());
+                Optional<String> explanation = Optional.of(tokens[1].substring(1, tokens[1].length() - 1));
+                response = new ServerResponse(ResponseType.CONTINUATION, Optional.empty(), explanation, new ArrayList<String>());
             } else {
                 lines.addLast(line);
             }
